refactor to directories by MVC
