--- __init__.py
@@ -1,11 +1,0 @@
-# from initialize.build import Build
-# from initialize.example_data import Example_data
-# from main import Main
-#
-# Build.connect()
-# Build.drop()
-# Build.create()
-# Example_data.insert()
-# Main.get_user_email_data()
-# Main.register()
-# Main.interview()
