+++ app.py
@@ -1,12 +1,12 @@
 from flask import *
-from read_from_text import *
+
+from controller.controll_admin import admin_page
+from initialize.build import Build
+from initialize.example_data import Example_data
+from initialize.read_from_text import *
 from main import Main
 from model.Applicant import Applicant
 from model.BaseModel import db
-from controll_admin import admin_page
-from initialize.build import Build
-from initialize.example_data import Example_data
-from main import Main
 
 app = Flask(__name__)
 app.register_blueprint(admin_page) # url_prefix='/admin'
@@ -18,12 +18,12 @@ secret = os.urandom(24)
 app.secret_key = secret
 
 Build.connect()
-Build.drop()
-Build.create()
-Example_data.insert()
-Main.get_user_email_data()
-Main.register()
-Main.interview()
+# Build.drop()
+# Build.create()
+# Example_data.insert()
+# Main.get_user_email_data()
+# Main.register()
+# Main.interview()
 
 @app.route('/')
 def index():
