+++ initialize/read_from_text.py
